Fix code doc linting errors in tests
