+++ plugins/woocommerce/tests/php/src/Internal/Admin/BlockTemplates/CustomBlockInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Automattic\WooCommerce\Tests\Internal\Admin\BlockTemplates;
+
+interface CustomBlockInterface {
+	/**
+	 * Custom method.
+	 */
+	public function custom_method(): string;
+}
